refactor: int -> long for phone

+++ app/src/main/java/com/obsessed/e_journal/Employee.java
@@ -3,7 +3,7 @@ package com.obsessed.e_journal;
 public class Employee extends Participant {
     private String position;
 
-    public Employee(String fullName, int phone, int cardID, String position) {
+    public Employee(String fullName, long phone, int cardID, String position) {
         super(fullName, phone, cardID);
         this.position = position;
     }
