+++ app/src/main/java/com/obsessed/e_journal/Parent.java
@@ -1,7 +1,7 @@
 package com.obsessed.e_journal;
 
 public class Parent extends Person {
-    public Parent(String fullName, int phone) {
+    public Parent(String fullName, long phone) {
         super(fullName, phone);
     }
 }
