+++ app/src/main/java/com/obsessed/e_journal/Participant.java
@@ -3,7 +3,7 @@ package com.obsessed.e_journal;
 public class Participant extends Person {
     private int cardID;
 
-    public Participant(String fullName, int phone, int cardID) {
+    public Participant(String fullName, long phone, int cardID) {
         super(fullName, phone);
         this.cardID = cardID;
     }
