+++ app/src/main/java/com/obsessed/e_journal/Person.java
@@ -2,9 +2,9 @@ package com.obsessed.e_journal;
 
 public class Person {
     private String fullName;
-    private int phone;
+    private long phone;
 
-    public Person(String fullName, int phone) {
+    public Person(String fullName, long phone) {
         this.fullName = fullName;
         this.phone = phone;
     }
