+++ app/src/main/java/com/obsessed/e_journal/Teacher.java
@@ -4,7 +4,7 @@ public class Teacher extends Participant{
     private String position;
     private String qualification;
 
-    public Teacher(String fullName, int phone, int cardID, String position, String qualification) {
+    public Teacher(String fullName, long phone, int cardID, String position, String qualification) {
         super(fullName, phone, cardID);
         this.position = position;
         this.qualification = qualification;
